requirements.txt
vectorbt
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0
plotly>=5.10.0
pyarrow>=8.0.0
ta>=0.10.0
pytest>=7.2.0
tqdm>=4.65.0
requests>=2.28.0
ccxt>=3.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tcl
python-dotenv
